teltonika-django/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
python-decouple==3.8
requests==2.31.0
pytz==2023.3 